Add package DotNetEnv
added package DotNetEnv to be able to create .env file that should
hold the infromation needed to connect to the database without exposing
it in the source code, nb the .env sould be added to the .gitignore
file.
Using the .env files will allow for having the same source code even if
we need diffrent passwords. §