add classes for queries and queries on roomTable
add query classes at the moment implementions for filtering on
diffrent facilities and to order by some collumns in asc/desc order is
possible.

to be implemented is filtering on price, rating, location
and bed amount/room sizes.

further more at the moment a filtering needs to be compleltly undone
if any Errors are commited.